# api/docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Тестовое задание",
        "title": "Test Task",
        "contact": {
            "name": "Evans Trein",
            "url": "https://github.com/EvansTrein",
            "email": "evanstrein@icloud.com"
        },
        "version": "0.1"
    },
    "host": "localhost:8000",
    "paths": {
        "/api/v1/wallet": {
            "post": {
                "description": "Perform a deposit or withdrawal operation on the wallet with the given walletID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Wallets"
                ],
                "summary": "Perform a wallet operation",
                "parameters": [
                    {
                        "description": "Wallet operation request",
                        "name": "wallet",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.WalletRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.RespMessage"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.RespError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.RespError"
                        }
                    }
                }
            }
        },
        "/api/v1/wallet/create": {
            "post": {
                "description": "Create a new wallet with zero balance and specified UUID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Wallets"
                ],
                "summary": "Create a new wallet",
                "parameters": [
                    {
                        "description": "Wallet object",
                        "name": "wallet",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.NewWallet"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.RespMessage"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.RespError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.RespError"
                        }
                    }
                }
            }
        },
        "/api/v1/wallets/{WALLET_UUID}": {
            "get": {
                "description": "Get the balance of the wallet with the given UUID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Wallets"
                ],
                "summary": "Get wallet balance",
                "parameters": [
                    {
                        "type": "string",
                        "example": "38c7e784-e963-4cc1-9124-de3e6c7e60e4",
                        "description": "UUID of the wallet",
                        "name": "WALLET_UUID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.ActiveWallet"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.RespError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.RespError"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.ActiveWallet": {
            "type": "object",
            "properties": {
                "total": {
                    "type": "number",
                    "example": 1000
                },
                "uuid": {
                    "type": "string",
                    "example": "38c7e784-e963-4cc1-9124-de3e6c7e60e4"
                }
            }
        },
        "models.NewWallet": {
            "type": "object",
            "required": [
                "walletId"
            ],
            "properties": {
                "walletId": {
                    "type": "string",
                    "example": "38c7e784-e963-4cc1-9124-de3e6c7e60e4"
                }
            }
        },
        "models.RespError": {
            "type": "object",
            "properties": {
                "messageErr": {
                    "type": "string",
                    "example": "user error"
                },
                "textErr": {
                    "type": "string",
                    "example": "error text"
                }
            }
        },
        "models.RespMessage": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "message"
                }
            }
        },
        "models.WalletRequest": {
            "type": "object",
            "required": [
                "amount",
                "operationType",
                "walletId"
            ],
            "properties": {
                "amount": {
                    "type": "number",
                    "example": 500
                },
                "operationType": {
                    "type": "string",
                    "example": "DEPOSIT or WITHDRAW"
                },
                "walletId": {
                    "type": "string",
                    "example": "38c7e784-e963-4cc1-9124-de3e6c7e60e4"
                }
            }
        }
    },
    "externalDocs": {
        "description": "OpenAPI",
        "url": "https://swagger.io/resources/open-api/"
    }
}